+++ packages/frontend/src/routes/(app)/category/[category=categories]/[slug]/+page.server.ts
@@ -58,9 +58,6 @@ export const load: PageServerLoad = (async (event: ServerLoadEvent) => {
 				// TODO
 			}
 		}
-		if (article.metaInfo.ogImage) {
-			// TODO
-		}
 
 		// Create an array of links to author's profile pages.
 		const ogAuthorLinks = [
