+++ packages/frontend/src/tests/helpers.test.ts
@@ -1,0 +1,12 @@
+import { hasUppercase } from '$lib/helpers';
+import { describe, expect, it } from 'vitest';
+
+describe('hasUppercase', () => {
+	it('is true for uppercase', () => {
+		expect(hasUppercase('abcDefg')).toBe(true);
+	});
+
+	it('is false for lowercase', () => {
+		expect(hasUppercase('abcdefg')).toBe(false);
+	});
+});
